Rename file mainInfo => MainInfo, because components should start with capital letter.
